requirements.txt
Django==3.2.13
django_allauth==0.44.0
djangorestframework==3.13.1
pandas==1.3.5
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
